requirements.txt
yfinance
pandas
psycopg2-binary
SQLAlchemy
pyyaml